+++ sign/db/db.ts
@@ -1,5 +1,6 @@
-import { CosmosClient, Container, ItemResponse } from '@azure/cosmos'
+import { CosmosClient, ItemResponse } from '@azure/cosmos'
 import { always } from 'ramda'
+import { createDBInstance } from './../../common/db/common'
 
 export type Secret = {
 	readonly id: string
@@ -9,28 +10,6 @@ export type Secret = {
 const SECRETS = {
 	database: 'Authentication',
 	container: 'Secrets',
-}
-
-const createDBInstance = async (
-	Client: typeof CosmosClient,
-	opts: {
-		readonly database: string
-		readonly container: string
-	},
-	env: NodeJS.ProcessEnv
-): Promise<Container> => {
-	const {
-		KHAOS_COSMOS_ENDPOINT: endpoint = '',
-		KHAOS_COSMOS_KEY: key = '',
-	} = env
-	const client = new Client({ endpoint, key })
-	const { database } = await client.databases.createIfNotExists({
-		id: opts.database,
-	})
-	const { container } = await database.containers.createIfNotExists({
-		id: opts.container,
-	})
-	return container
 }
 
 export const writer = (client: typeof CosmosClient) => async (
